Merges changes from upstream.
Adds new functionality for !maybe_kick (1 in 10 change of kicking).

Adds new functionality to reload the aliases live.

Adds a check for nick password - only tries to auth if not undefined in
config.

Some updates to the aliases object (which will be replaced shortly with
an actual database);

(Shitty commit message and commit in general)